Extract AI logic in ai.js file.
